% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_reduc.R
\name{filter_reduc}
\alias{filter_reduc}
\title{filter_reduc}
\usage{
filter_reduc(
  df = filters_neg_up,
  filtre = NULL,
  obj = NULL,
  direction = "below",
  param_increase = NULL,
  param_reduce = NULL,
  arrangeDf = T
)
}
\description{
filter_reduc
}
