% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_relative_param.R
\name{find_relative}
\alias{find_relative}
\title{Title}
\usage{
find_relative(
  ...,
  model = model_RxODE,
  time_simul = times,
  states = initial_cmt_values,
  values = domain,
  protocol = "dose50",
  sensitivity = 1e-06,
  printPlot = T,
  deepAnalysis = F,
  printCode = T
)
}
\value{

}
\description{
Title
}
\examples{

}
