% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_algo2.R
\docType{data}
\name{npersalve}
\alias{npersalve}
\title{Title}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
npersalve
}
\arguments{
\item{add_events}{}
}
\value{

}
\description{
Title
}
\examples{

}
\keyword{datasets}
