% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{simulations}
\alias{simulations}
\title{Title}
\usage{
simulations(
  ind_param = double(),
  add_events = tibble(),
  returnSim = T,
  icmt = initial_cmt_values,
  time_vec = times,
  pardf = parameters_default_values,
  model = model_RxODE
)
}
\arguments{
\item{add_events}{}
}
\value{

}
\description{
Title
}
\examples{

}
