% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_R6_creation.R
\name{VP_proj_creator}
\alias{VP_proj_creator}
\title{Initiate a R6 object project}
\description{
Initiate a R6 object project

Initiate a R6 object project
}
\author{
Thibaud Derippe
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VP_proj_creator$new()}}
\item \href{#method-print}{\code{VP_proj_creator$print()}}
\item \href{#method-add_VP}{\code{VP_proj_creator$add_VP()}}
\item \href{#method-set_targets}{\code{VP_proj_creator$set_targets()}}
\item \href{#method-algo2}{\code{VP_proj_creator$algo2()}}
\item \href{#method-big_screening}{\code{VP_proj_creator$big_screening()}}
\item \href{#method-compute_zone_maybe}{\code{VP_proj_creator$compute_zone_maybe()}}
\item \href{#method-compute_zone_sure}{\code{VP_proj_creator$compute_zone_sure()}}
\item \href{#method-fill_simul}{\code{VP_proj_creator$fill_simul()}}
\item \href{#method-launch_programmed}{\code{VP_proj_creator$launch_programmed()}}
\item \href{#method-make_filters}{\code{VP_proj_creator$make_filters()}}
\item \href{#method-n_filter_reduc}{\code{VP_proj_creator$n_filter_reduc()}}
\item \href{#method-plot_2D}{\code{VP_proj_creator$plot_2D()}}
\item \href{#method-plot_3D}{\code{VP_proj_creator$plot_3D()}}
\item \href{#method-plot_VP}{\code{VP_proj_creator$plot_VP()}}
\item \href{#method-time_analyserr}{\code{VP_proj_creator$time_analyserr()}}
\item \href{#method-clone}{\code{VP_proj_creator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$new(
  sourcefile = file.path(find.package("QSPVP"), "config_Simeoni.r")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_VP"></a>}}
\if{latex}{\out{\hypertarget{method-add_VP}{}}}
\subsection{Method \code{add_VP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$add_VP(
  VP_df,
  fix_df = NULL,
  saven = 50,
  drug = NULL,
  update_at_end = T,
  time_compteur = F,
  fillatend = F,
  reducefilteratend = F,
  npersalve = 1000,
  use_green_filter = F,
  pctActivGreen = 0.1,
  keepRedFiltaftDis = T,
  methodFilter = 2,
  use_red_filter = T,
  cmtalwaysbelow = NULL,
  keep = NULL,
  saveVPRej = T,
  timeSave = NULL,
  PreviousResults = tibble(),
  RedFilterDisAllProt = F,
  GreenFilterDisAllProt = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_targets"></a>}}
\if{latex}{\out{\hypertarget{method-set_targets}{}}}
\subsection{Method \code{set_targets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$set_targets(
  ...,
  filter = NULL,
  ntime = 3,
  manual = NULL,
  timeforce = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-algo2"></a>}}
\if{latex}{\out{\hypertarget{method-algo2}{}}}
\subsection{Method \code{algo2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$algo2(
  domain,
  fix = NULL,
  npersalve = 2e+05,
  npersalveFinal = 1e+06,
  file = "",
  save_every = 5,
  method = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-big_screening"></a>}}
\if{latex}{\out{\hypertarget{method-big_screening}{}}}
\subsection{Method \code{big_screening()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$big_screening(domain)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_zone_maybe"></a>}}
\if{latex}{\out{\hypertarget{method-compute_zone_maybe}{}}}
\subsection{Method \code{compute_zone_maybe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$compute_zone_maybe(
  zone_sure = F,
  keptSingleValue = T,
  limits = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_zone_sure"></a>}}
\if{latex}{\out{\hypertarget{method-compute_zone_sure}{}}}
\subsection{Method \code{compute_zone_sure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$compute_zone_sure(domain, keptSingleValue = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_simul"></a>}}
\if{latex}{\out{\hypertarget{method-fill_simul}{}}}
\subsection{Method \code{fill_simul()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$fill_simul(nsalve = 500)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch_programmed"></a>}}
\if{latex}{\out{\hypertarget{method-launch_programmed}{}}}
\subsection{Method \code{launch_programmed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$launch_programmed()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_filters"></a>}}
\if{latex}{\out{\hypertarget{method-make_filters}{}}}
\subsection{Method \code{make_filters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$make_filters(cmt = "tumVol")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-n_filter_reduc"></a>}}
\if{latex}{\out{\hypertarget{method-n_filter_reduc}{}}}
\subsection{Method \code{n_filter_reduc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$n_filter_reduc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_2D"></a>}}
\if{latex}{\out{\hypertarget{method-plot_2D}{}}}
\subsection{Method \code{plot_2D()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$plot_2D(
  x,
  y,
  cmt_green = "tumVol",
  toaddneg = NULL,
  plotMain = F,
  add_point = F,
  IDref = NULL,
  plotoreturn = 3
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_3D"></a>}}
\if{latex}{\out{\hypertarget{method-plot_3D}{}}}
\subsection{Method \code{plot_3D()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$plot_3D(x, y, z, toaddneg = NULL, add_point = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_VP"></a>}}
\if{latex}{\out{\hypertarget{method-plot_VP}{}}}
\subsection{Method \code{plot_VP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$plot_VP(nmax = Inf, rowids = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-time_analyserr"></a>}}
\if{latex}{\out{\hypertarget{method-time_analyserr}{}}}
\subsection{Method \code{time_analyserr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$time_analyserr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VP_proj_creator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
